Add helpers file in the backend
