+++ src/server/helpers.js
@@ -1,0 +1,19 @@
+const getToday = () => {
+  let today = new Date();
+  let dd = today.getDate();
+  let mm = today.getMonth() + 1; //January is 0!
+
+  let yyyy = today.getFullYear();
+  if (dd < 10) {
+    dd = `0${dd}`;
+  }
+  if (mm < 10) {
+    mm = `0${mm}`;
+  }
+  today = `${yyyy}-${mm}-${dd}`;
+  return today;
+};
+
+module.exports = {
+  getToday: getToday
+};
